+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.1.2] - Ongoing
+## [2.1.2] - 2024-09-27
 
 This update introduces bug fixes
 
 ### Added
  - Add getter methods for installation lifecycle steps
+ - Added pre-install & post-install execution unit handling
 
 ### Changed
  - Updated DocStrings in PackageManager